WIP: begin support for vector-valued objectives
To be read in conjunction with https://github.com/jump-dev/MathOptInterface.jl/pull/2070

A proof-of-concept solver and JuMP examples are available at https://github.com/odow/MOO.jl/pull/2

The issue https://github.com/jump-dev/JuMP.jl/issues/2099 discussed two approaches
for implementing MO in JuMP.

 1. Treat multicriteria as a vector of scalar objectives and a vector of
    scalar senses. This would let someone write [Min, Max], [f(x), g(x)].
    The main reason for this approach is that it matches what users want
    to do.
 2. Treat multicriteria as an optimization problem with a vector-valued
    objective function. Users could write only Min, f(x) where f(x) is a
    vector. The main reason for this approach is that it matches what
    MathOptInterface wants.

This PR implements option 2. The strongest reason in support of option 2
is that it requires very little code to implement, suggesting that it is
a natural extension of MOI.

The biggest downside is that it doesn't overcome the Min-Max issue; but
I think we can work around this with user-facing cosmetic tooling in JuMP;
solvers would be forced to accept a single sense.